Separated pick manager from vs
